backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# PML parsing
procmon-parser==0.3.9

# Data handling
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# AI integration
openai==1.58.0
httpx==0.28.0

# PDF generation
reportlab==4.2.5
weasyprint==63.0

# YAML for rules
pyyaml==6.0.2

# Async support
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0

# Logging
structlog==24.4.0
